+++ changelog.md
@@ -10,6 +10,7 @@ TGUI 0.9.4  (TBD)
 - ListBox::setMaximumItems didn't reset selected item when removed
 - Fixed issue with renderer property within custom renderer outside tgui namespace
 - Widget could remain in hover state after hiding it
+- Fixed broken text rendering in SDL/OpenGL backend
 
 
 TGUI 0.9.3  (15 January 2022)